+++ AzureCP/AzureCP.cs
@@ -1159,7 +1159,9 @@ namespace azurecp
                 }
                 else
                 {
-                    currentFilter = String.Format(preferredFilterPattern, currentPropertyString);
+                    // Use String.Replace instead of String.Format because String.Format trows an exception if input contains a '{'
+                    //currentFilter = String.Format(preferredFilterPattern, currentPropertyString);
+                    currentFilter = preferredFilterPattern.Replace("{0}", currentPropertyString);
                 }
 
                 // Id needs a specific check: input must be a valid GUID AND equals filter must be used, otherwise Azure AD will throw an error